requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
Django==5.0.3
django-cors-headers==4.3.1
django-mail-templated==2.6.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.51.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.4
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==8.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
